move emacs mode for zle before sourcing *.zsh
My $EDITOR is not being set early enough for zsh to set the zle mode (to
vi). So in a top level shell, 'bindkey -e' did nothing because the mode
was already set. However, in any nested shell (like tmux), $EDITOR is set
ahead of time and zsh sest the zle mode to vi. This causes any key
bindings set before 'bindkey -e' to get overridden. My solution is
currently to move 'bindkey -e' to .zshrc because the zsh directory will
get sourced last out of the topic directories.

In the future, this should likely get taken out of .zshrc however it
will likely require a larger rewrite.