fix programs not having access to XDG vars
when I added XDG vars, I did not export them in zshrc. Any application
trying to use one of these variables was then unable to. This was first
discovered in Vim, since I just added VIMINIT and moved coniguration into
XDG_CONFIG_HOME. This issue was not noticed initially because I have a
.localrc on one of my devices that exports them properly.