remove gdircolors conditional, BSD LS config
Since BSD coreutils are replaced with GNU coreutils,
configuration specific to BSD ls is no longer needed (CLICOLOR and
LSCOLORS). In addition, checking for gdircolors is no longer needed
because dircolors will refer to the GNU version by default.

In addition, the output of the dircolors command was put in
dircolors.zsh so that it doesn't require an eval.