add dotfile utils, remove some holman based setup
Holman's dotfiles setup is topic based, where a folder is for an
applicaton and it can contain path/completion/*.zsh files that are
loaded in a specific order.

I've found that I rarely use this separation, so I'm migrating away from
it. At this point, there are no path.zsh files and this commit removes
the (only) completion.zsh, so the separate load points are removed as
well.