fix: vim $PATH empty inside tmux
My path_helper 'bad hack' was causing shells inside vim to have an empty
path variable when vim was used inside tmux. To remedy this, VIMINIT has
been updated to set the MYVIMRC variable. This variable is then used as a
conditional to prevent clearing the path because shells in vim don't
trigger the path_helper issue