feat: use tm/popup for switching sessions in tmux
Replaced the built in session switcher in tmux with the tm script, which
allows for fuzzy session switching. This was previously difficult to do
with a keybind because it required creating a temporary pane/window/etc.
tmux 3.2 introduced popup windows, and the fzf-tmux script makes it easy
to put fzf inside one, so tm was updated to use it.

tm also had to be moved from a zsh function to a bash script so that it
can be called by the tmux popup. The popup executes as a non-interactive
shell so it doesn't have its fpath updated to include my functions.
However, it does inherit $PATH, so a shell script in $DOTFILES/bin works
just fine.

There's also an unrelated change in tmux.conf that was recommended by
neovim's :checkhealth that allows it to detect focus changes inside
tmux.