move prompt configuration into .config
zsh/plugins/echo.zsh was a file leftover when I was using the Pure
prompt, which requires adding files to fpath but not sourcing them. Its
purpose was to make it very obvious if I was sourcing the plugins
folder by mistake. Since that is no longer a requirement, the echo is
no longer needed.