fix tm switching session on ESC
When fzf didn't return a session to switch to, tm was running:

tmux $change -t ''

which was jumping to the previous session instead of erroring (it errors
if you try to run that command in a terminal).

This fixes that case, as well as removing a shellcheck error and
enabling -euo pipefail. A comment is left describing why shellcheck
SC2016 is ignored, and the rest of the file doesn't have any other
errors.

- checking for presence of $1 was improved by checking the length
of arguments instead of "$1" being blank.
- checking for presence of $TMUX was improved, as now it will not throw
  an error in 'set -u' when the variable is unbound. The + parameter
  expansion will either evaluate to "" when $TMUX is unset or "x" when
  it is set. Since -n was changed to -z, the change assignment was also
  flipped.