fix bootstrap script not being idempotent
The bug was introduced in 2c61bd2 when trying to fix an unbound variable
error for $skip. However, the parameter expansion used is actually
incorrect. ${skip+false} will return "false" depending on whether $skip
is defined, but it should have been ${skip-false} which will return
"false" ONLY when $skip is unbound.

Instead of fixing the parameter expansion, I gave $skip a value at the
beginning of the function so that it isn't unbound. This seems like a
much easier to understand solution and doesn't depend on knowing how
parameter expansion works too much.