really fix idempotency of bootstrap script
The last idempotency fix (3879d0c) did ensure that the symlinking part
of the script could be run multiple times, but I missed that the script
was exiting with an error code on the test for vim-plug being installed
(because of set -e).

This time I actually read the man page of the "test" command and found
that I could use -e in the test which returns whether the file exists
(but makes no assertions about whether it is a file or directory). This
allowed the existence check to be moved inside prompt_install, fixing
the early exit issue and making the code much cleaner.