Add $HOME/.local/bin to path
I've been building edge versions of tools (tmux, neovim, and now sdev)
more often, and it's lead to me wondering what the best place to put
those binaries is. My understanding now is that Bash, systemd-path, and
XDG all seem to agree that $HOME/.local/bin is the place. It's being
added to zshrc because Zsh does not include this directory in the path
by default (Bash does).

Since $HOME/.local/bin is controlled by the user, it should almost
certainly be at the front of the path. To ensure nothing ends up before
it, the prepending is almost last in the zshrc. $DOTFILES/bin is being
put there as well both for organization and because it makes sense that
things in my dotfiles are likely wanted over asdf shims which were
previously proceeding it.