Fix tm partially matching explicit session names
Today I learned that tmux matches target-session in four stages,
including exact name and as a prefix of session names. So if I created a
'dotfiles' session, then `$ tm dot` would attach to it instead of
creating a new session named 'dot'.

The intended behavior has always been just `$ tm` to fuzzy attach to
sessions and `$ tm dot` to create/attach to 'dot' exactly.

This is done in tmux by prefixing the target-session with a '='