Make tm's fzf preview work on BSD sed
Previously it was using -z to process the whole file at once instead of
line by line (so the regex could easily select all the \n at the end of
the file). However, -z is only supported by GNU sed.

I ended up finding this snippet in a stackoverflow answer:
https://stackoverflow.com/questions/7359527/removing-trailing-starting-newlines-with-sed-awk-tr-and-friends

It claims that http://sed.sourceforge.net/sed1line.txt is the source of
the snippet but the examples I see there are slightly different:

  # delete all trailing blank lines at end of file
  sed -e :a -e '/^\n*$/{$d;N;ba' -e '}'  # works on all seds
  sed -e :a -e '/^\n*$/N;/\n$/ba'        # ditto, except for gsed 3.02.*

The one from stack is working and the ones from sed1line are not, but I
have not really dug into it much