Add :Coauthor function for fuzzy searching authors
I haven't experimented very much with Vim configuration, so this is
probably a little rough. The general idea is that this will make it
easier to add Co-authored-by lines to commits by fuzzy searching anyone
who has previously committed to the current repository. Obviously, this
requires the person to have committed to the repository before so it
isn't perfect, but it should help me not have to go looking up old
commits to guess.