Fix using RGB colors in tmux with alacritty
When installed using Brew, Alacritty's terminfo is correctly installed
and $TERM ends up being set to alacritty (instead of xterm-256color).
Currently, my tmux configuration only sets the RGB override for
terminfos that start with xterm, so the alacritty terminfo must also be
added to the list of terminfos to override.