Make glum more dynamic -> glu
glum has generally been a really useful alias, but not all repositories
used "main" as their default branch.

To make it more useful for all repositories, this commit replaces "main"
with "$(git branch --show-current)". This enables glu (renamed since
"main" is not longer hard coded) to work with repositories that use
other default branches, and more generally it can be used to sync any
branch on a remote that "tracks" an upstream.