Fix tm outside of tmux
When tm was updated to use `sdev`, it would unconditionally display the
picker inside `tmux display-popup`. However, the popup is only available
inside tmux so it would error when run outside tmux.

This commit fixes tm outside of tmux by checking whether tm is being run
inside tmux, and only using the popup when it is.